requirements.txt
flask==2.1.0
flask-cors==3.0.10
google-generativeai==0.3.1
python-dotenv==0.20.0
requests==2.27.1
beautifulsoup4==4.11.1